Offer manual option of booting Agent in thread of user choice
